\docType{data}
\name{deb_complete}
\alias{deb_complete}
\title{Deb Roy Fake Lab Motion & Dialogue (Long Format)}
\format{A data frame with 120 rows and 6 variables}
\usage{
data(deb_complete)
}
\description{
A fictional long dataset containing fabricated motion paths
and dialogue for Deb Roy, Philip DeCamp, Brandon C. Roy,
and Rony Kubat in Deb's MIT lab.
}
\details{
\itemize{ \item id. The image frame id number.  \item time.
The video time.  \item person. The person who is moving.
\item coord. The grid location of the individual at that
time.  \item dialogue. The speech being used (taken from
Romeo and Juliet.  \item wc. Word count.  \item x. An x
coordinate (developed from \code{gridify}).  \item y. A y
coordinate (developed from \code{gridify}). }
}
\references{
\url{http://www.ted.com/talks/deb_roy_the_birth_of_a_word.html}
}
\keyword{datasets}

