\docType{data}
\name{deb_dialogue}
\alias{deb_dialogue}
\title{Deb Roy Fake Dialogue}
\format{A data frame with 30 rows and 3 variables}
\usage{
data(deb_dialogue)
}
\description{
A fictional dataset containing fabricated dialogue for Deb
Roy, Philip DeCamp, Brandon C. Roy, and Rony Kubat in Deb's
MIT lab.  Dialogue taken from Romeo and Juliet.
}
\details{
\itemize{ \item person. The person who is talking.  \item
time. The video time.  \item dialogue. The speech being
used (taken from Romeo and Juliet. }
}
\keyword{datasets}

