\docType{data}
\name{deb_long}
\alias{deb_long}
\title{Deb Roy Ted Talk Long}
\format{A data frame with 120 rows and 4 variables}
\usage{
data(deb_long)
}
\description{
A fictional dataset containing fabricated motion paths for
Deb Roy, Philip DeCamp, Brandon C. Roy, and Rony Kubat in
Deb's MIT lab.
}
\details{
\itemize{ \item id. The image frame id number.  \item time.
The video time.  \item person. The person who is moving.
\item coord. The grid location of the individual at that
time. }
}
\references{
\url{http://www.ted.com/talks/deb_roy_the_birth_of_a_word.html}
}
\keyword{datasets}

