\docType{data}
\name{deb_wide}
\alias{deb_wide}
\title{Deb Roy Ted Talk Wide}
\format{A data frame with 30 rows and 6 variables}
\usage{
data(deb_wide)
}
\description{
A fictional dataset containing fabricated motion paths for
Deb Roy, Philip DeCamp, Brandon C. Roy, and Rony Kubat in
Deb's MIT lab.
}
\details{
\itemize{ \item id. The image frame id number.  \item time.
The video time.  \item deb. Fake movement for Deb Roy.
\item philip. Fake movement for Philip DeCamp.  \item rony.
Fake movement for Rony Kubat.  \item brandon. Fake movement
for Brandon C. Roy. }
}
\references{
\url{http://www.ted.com/talks/deb_roy_the_birth_of_a_word.html}
}
\keyword{datasets}

