\name{merge_pdf}
\alias{merge_pdf}
\title{Merge Multiple PDFs}
\usage{
merge_pdf(pdfs, out)
}
\arguments{
  \item{pdfs}{A vector of paths to multiple PDFs.}

  \item{out}{Path to the output merged PDF.}
}
\value{
Returns a single combined plot.
}
\description{
Allows for merging of multiple PDF files.
}
\author{
Ananda Mahto and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://trinkerrstuff.wordpress.com/2012/10/08/splitting-and-combining-r-pdf-visuals/}
}

