\name{plotgrid}
\alias{plotgrid}
\title{Title}
\usage{
plotgrid(file, coord.color = "gray60", grid.color = coord.color,
  text.size = 3, grid.data = grid_calc())
}
\arguments{
  \item{file}{The imput png file.}

  \item{coord.color}{The color to make the coordinate
  labels.}

  \item{grid.color}{The color to make the grid.}

  \item{text.size}{The size of the coordinate labels.}

  \item{grid.data}{The output from the \code{grid_calc}
  function.}
}
\value{
Returns an image with a grid and labels.
}
\description{
Description
}
\author{
Troy (\url{stackoverflow.com}) and Tyler Rinker
<tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/21043587/1000343}
}

