\name{mp4_duration}
\alias{mp4_duration}
\title{Video Duration}
\usage{
mp4_duration(path)
}
\arguments{
  \item{path}{Path to the in .mp4 file.}
}
\value{
A numeric value giving length of video in seconds.
}
\description{
Reports the length of a video in seconds.
}
\examples{
\dontrun{
mp4_duration("foo.mp4")
}
}
\keyword{duration}

